fixed merged conflicts and refactored view settings
