+++ view/GameView/GameView.py
@@ -2,6 +2,7 @@ import pygame
 
 from view.BasicView import BasicView
 from view.GameView.GameViewController import GameViewController
+from view.ViewSettings import ViewSettings
 from controller.ControllerView import ControllerGameView
 
 
@@ -9,11 +10,11 @@ class GameView(BasicView):
     """
     This class implements the interface of the GameView classes to the main controller
     """
-    def __init__(self, window: pygame.display, controller: ControllerGameView) -> None:
-        super().__init__(window, controller)
-        self.window_width, self.window_height = pygame.display.get_surface().get_size()
-        self.gameViewController = GameViewController(self)
 
+    def __init__(self, window: pygame.display, controller: ControllerGameView, setttings: ViewSettings) -> None:
+        super().__init__(window, controller, setttings)
+        self.settings = setttings
+        self.gameViewController = GameViewController(self, setttings)
 
     def draw(self) -> None:
         self.window.fill((50, 50, 50))
@@ -23,6 +24,6 @@ class GameView(BasicView):
         pygame.display.flip()
 
     def receive_event(self, event: pygame.event.Event) -> None:
-        #todo for now all events are forwarded
-        #here the filtering between HUD and Game should take place!
+        # todo for now all events are forwarded
+        # here the filtering between HUD and Game should take place!
         self.gameViewController.receive_event(event)
